Step 3-4
Open WXController.m and replace the boilerplate -viewDidLoad method with
the following:

    - (void)viewDidLoad {
        [super viewDidLoad];

        // Remove this later
        self.view.backgroundColor = [UIColor redColor];
    }

The status bar is a little difficult to read against the red background.
Fortunately, there’s an easy way to make the status bar a lot more
legible.

There’s a new API in UIViewController in iOS 7 to control the appearance
of the status bar. Open WXController and add the following code directly
below -viewDidLoad:

    - (UIStatusBarStyle)preferredStatusBarStyle {
        return UIStatusBarStyleLightContent;
    }