Step 4-6
Create a tableview that will handle all the data presentation.
WXController will be the delegate and data source, as well as the scroll
view delegate. Note that you’re also setting pagingEnabled to YES.

NOTE: the scroll view stops on multiples of the scroll view’s bounds
when the user scrolls. The default value is NO.