Step 5-2
Creating Your Weather Model

Your models will use Mantle which makes data mapping and value
transformation a cinch.

The MTLJSONSerializing protocol tells the Mantle serializer that this
object has instructions on how to map JSON to Objective-C properties.

Properties are all of your weather data properties. You’ll be using a
couple of them, but its nice to have access to all of the data in the
event that you want to extend your app down the road.

imageName is simply a helper method to map weather conditions to imageName
files.